backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
boto3>=1.18.0
aiocache>=0.11.1
passlib>=1.7.4
python-jose>=3.3.0
redis>=4.0.0
pydantic>=1.8.2
bcrypt>=4.0.0
python-dotenv>=0.19.0
pandas>=1.5.0
python-multipart>=0.0.5
openpyxl>=3.0.0
